# apps/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OpportunityCard": {
        "description": "Compact representation for the opportunity feed.",
        "properties": {
          "audience_fit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audience Fit"
          },
          "campaign_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Campaign Id"
          },
          "classification": {
            "title": "Classification",
            "type": "string"
          },
          "commercial_value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Commercial Value"
          },
          "confidence_level": {
            "title": "Confidence Level",
            "type": "string"
          },
          "confidence_score": {
            "title": "Confidence Score",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "decision": {
            "title": "Decision",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "inside_scout_area": {
            "title": "Inside Scout Area",
            "type": "boolean"
          },
          "is_simulated": {
            "title": "Is Simulated",
            "type": "boolean"
          },
          "location_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Id"
          },
          "opportunity_score": {
            "title": "Opportunity Score",
            "type": "integer"
          },
          "priority_score": {
            "title": "Priority Score",
            "type": "integer"
          },
          "recommended_action": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recommended Action"
          },
          "relevance_score": {
            "title": "Relevance Score",
            "type": "integer"
          },
          "resolved_market": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved Market"
          },
          "risk_level": {
            "title": "Risk Level",
            "type": "string"
          },
          "scout_request_id": {
            "title": "Scout Request Id",
            "type": "string"
          },
          "source_summary": {
            "items": {
              "type": "string"
            },
            "title": "Source Summary",
            "type": "array"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "urgency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Urgency"
          },
          "why_it_matters": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Why It Matters"
          }
        },
        "required": [
          "id",
          "title",
          "classification",
          "decision",
          "opportunity_score",
          "confidence_score",
          "confidence_level",
          "priority_score",
          "relevance_score",
          "risk_level",
          "resolved_market",
          "inside_scout_area",
          "why_it_matters",
          "recommended_action",
          "audience_fit",
          "urgency",
          "commercial_value",
          "source_summary",
          "status",
          "location_id",
          "campaign_id",
          "scout_request_id",
          "is_simulated",
          "created_at"
        ],
        "title": "OpportunityCard",
        "type": "object"
      },
      "OpportunityDetail": {
        "properties": {
          "ai_inference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Inference"
          },
          "audience_fit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audience Fit"
          },
          "brand_id": {
            "title": "Brand Id",
            "type": "string"
          },
          "campaign_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Campaign Id"
          },
          "claims_warnings": {
            "items": {
              "type": "string"
            },
            "title": "Claims Warnings",
            "type": "array"
          },
          "classification": {
            "title": "Classification",
            "type": "string"
          },
          "commercial_value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Commercial Value"
          },
          "confidence_level": {
            "title": "Confidence Level",
            "type": "string"
          },
          "confidence_score": {
            "title": "Confidence Score",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "decision": {
            "title": "Decision",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "inside_scout_area": {
            "title": "Inside Scout Area",
            "type": "boolean"
          },
          "is_simulated": {
            "title": "Is Simulated",
            "type": "boolean"
          },
          "location_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Id"
          },
          "observed_evidence": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Observed Evidence",
            "type": "array"
          },
          "opportunity_score": {
            "title": "Opportunity Score",
            "type": "integer"
          },
          "priority_score": {
            "title": "Priority Score",
            "type": "integer"
          },
          "recommended_action": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recommended Action"
          },
          "relevance_score": {
            "title": "Relevance Score",
            "type": "integer"
          },
          "resolved_market": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved Market"
          },
          "risk_level": {
            "title": "Risk Level",
            "type": "string"
          },
          "risk_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Note"
          },
          "scores": {
            "items": {
              "$ref": "#/components/schemas/ScoreBreakdown"
            },
            "title": "Scores",
            "type": "array"
          },
          "scout_request_id": {
            "title": "Scout Request Id",
            "type": "string"
          },
          "source_summary": {
            "items": {
              "type": "string"
            },
            "title": "Source Summary",
            "type": "array"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "suggested_angles": {
            "items": {
              "type": "string"
            },
            "title": "Suggested Angles",
            "type": "array"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "urgency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Urgency"
          },
          "validation_evidence": {
            "items": {
              "$ref": "#/components/schemas/ValidationEvidenceOut"
            },
            "title": "Validation Evidence",
            "type": "array"
          },
          "who_cares": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Who Cares"
          },
          "why_it_matters": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Why It Matters"
          }
        },
        "required": [
          "id",
          "title",
          "classification",
          "decision",
          "opportunity_score",
          "confidence_score",
          "confidence_level",
          "priority_score",
          "relevance_score",
          "risk_level",
          "resolved_market",
          "inside_scout_area",
          "why_it_matters",
          "recommended_action",
          "audience_fit",
          "urgency",
          "commercial_value",
          "source_summary",
          "status",
          "location_id",
          "campaign_id",
          "scout_request_id",
          "is_simulated",
          "created_at",
          "who_cares",
          "observed_evidence",
          "ai_inference",
          "suggested_angles",
          "risk_note",
          "claims_warnings",
          "brand_id"
        ],
        "title": "OpportunityDetail",
        "type": "object"
      },
      "OpportunityStatusUpdate": {
        "properties": {
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "OpportunityStatusUpdate",
        "type": "object"
      },
...
      "ScoreBreakdown": {
        "properties": {
          "breakdown": {
            "additionalProperties": true,
            "title": "Breakdown",
            "type": "object"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "kind",
          "total",
          "breakdown"
        ],
        "title": "ScoreBreakdown",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "ValidationEvidenceOut": {
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "source_type": {
            "title": "Source Type",
            "type": "string"
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          },
          "weight": {
            "title": "Weight",
            "type": "number"
          }
        },
        "required": [
          "source_type",
          "detail",
          "weight"
        ],
        "title": "ValidationEvidenceOut",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "SignalNest API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/workspaces/{workspace_id}/opportunities": {
      "get": {
        "operationId": "list_opportunities_api_v1_workspaces__workspace_id__opportunities_get",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "location_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Location Id"
            }
          },
          {
            "in": "query",
            "name": "campaign_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Campaign Id"
            }
          },
          {
            "in": "query",
            "name": "scout_request_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scout Request Id"
            }
          },
          {
            "in": "query",
            "name": "classification",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Classification"
            }
          },
          {
            "in": "query",
            "name": "decision",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Decision"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "risk_level",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Risk Level"
            }
          },
          {
            "in": "query",
            "name": "market",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Market"
            }
          },
          {
            "in": "query",
            "name": "min_score",
            "required": false,
            "schema": {
              "default": 0,
              "maximum": 100,
              "minimum": 0,
              "title": "Min Score",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "priority",
              "title": "Sort",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "order",
            "required": false,
            "schema": {
              "default": "desc",
              "pattern": "^(asc|desc)$",
              "title": "Order",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/OpportunityCard"
                  },
                  "title": "Response List Opportunities Api V1 Workspaces  Workspace Id  Opportunities Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Opportunities",
        "tags": [
          "opportunities"
        ]
      }
    },
    "/api/v1/workspaces/{workspace_id}/opportunities/{opportunity_id}": {
      "get": {
        "operationId": "get_opportunity_api_v1_workspaces__workspace_id__opportunities__opportunity_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "opportunity_id",
            "required": true,
            "schema": {
              "title": "Opportunity Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpportunityDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Opportunity",
        "tags": [
          "opportunities"
        ]
      }
    },
    "/api/v1/workspaces/{workspace_id}/opportunities/{opportunity_id}/status": {
      "put": {
        "operationId": "update_opportunity_status_api_v1_workspaces__workspace_id__opportunities__opportunity_id__status_put",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "opportunity_id",
            "required": true,
            "schema": {
              "title": "Opportunity Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpportunityStatusUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpportunityCard"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Opportunity Status",
        "tags": [
          "opportunities"
        ]
      }
    },
...
  }
}